# api/openapi.yaml
openapi: 3.1.0
info:
  title: Vermouth gateway
  version: 0.1.0
  description: |
    The single description of the gateway surface (STK-10). The gateway's Go
    request and response types are generated from this document with
    oapi-codegen, and the browser's types with openapi-typescript. A hand
    written client type for a gateway endpoint is not allowed.

    Only the gateway is described here. Services are reached by the gateway
    alone (INV-1), so their endpoints are not a public surface.

    The four auth endpoints are unauthenticated because they are what produces
    a token. Everything else under /api/ needs one. The teaching endpoints are
    the first real product thread through the gateway.
servers:
  - url: http://localhost:8080
    description: local development
paths:
...
  /api/classes/{class_id}/rates:
    get:
      operationId: getClassRates
      summary: Read authoritative current rate and projected dated history
      parameters:
        - $ref: "#/components/parameters/ClassId"
      responses:
        "200":
          description: Current rate, allowed correction range, and complete dated history
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRates"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
...
  /api/invoice-profile:
    get:
      operationId: getInvoiceProfile
      summary: Read the signed in tutor's invoice identity and bank details
      responses:
        "200":
          description: The canonical private invoice profile
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceProfile"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: putInvoiceProfile
      summary: Save the signed in tutor's complete editable invoice profile representation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutInvoiceProfileRequest"
      responses:
        "200":
          description: The normalized canonical profile after the save
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/banks:
    get:
      operationId: getBanks
      summary: Read the active committed Vietnamese bank catalog
      responses:
        "200":
          description: Every active catalog entry in stable payment code order
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: "private, max-age=86400"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankCatalog"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/billing-periods/default:
    get:
      operationId: getBillingPeriodDefault
      summary: Read the server derived most recent completed month
      responses:
        "200":
          description: Verified timezone and default completed period
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPeriodDefault"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/billing-periods/{year}/{month}:
    get:
      operationId: getBillingPeriod
      summary: Read an unissued month or its immutable issued run
      parameters:
        - $ref: "#/components/parameters/BillingYear"
        - $ref: "#/components/parameters/BillingMonth"
      responses:
        "200":
          description: The current authoritative period state
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPeriodState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/billing-periods/{year}/{month}/preview:
    post:
      operationId: previewBillingPeriod
      summary: Calculate one completed month after proving projection readiness
      parameters:
        - $ref: "#/components/parameters/BillingYear"
        - $ref: "#/components/parameters/BillingMonth"
      responses:
        "200":
          description: Ready, blocked, empty, or already issued month
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingPreview"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ProjectionUnavailable"
  /api/billing-periods/{year}/{month}/issue:
    post:
      operationId: issueBillingPeriod
      summary: Issue one immutable invoice per billable student
      parameters:
        - $ref: "#/components/parameters/BillingYear"
        - $ref: "#/components/parameters/BillingMonth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueBillingRequest"
      responses:
        "200":
          description: The existing run returned for a retry
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRun"
        "201":
          description: The new run and every immutable invoice
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingRun"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ProjectionUnavailable"
...
components:
  parameters:
...
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BillingYear:
      name: year
      in: path
      required: true
      schema:
        type: integer
        minimum: 2000
    BillingMonth:
      name: month
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 12
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An Ed25519 signed access token, verified locally by each service (INV-14).
  responses:
    BadRequest:
      description: The request was not acceptable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: No valid token was supplied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: No owned resource matched
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The requested state conflicts with committed state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: The request shape is valid but one or more profile fields are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: The owning service could not complete the request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: A required service did not answer successfully
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ProjectionUnavailable:
      description: The projection cannot yet prove the requested published cut
      headers:
        Retry-After:
          required: false
          schema:
            type: integer
            minimum: 1
        Cache-Control:
          required: true
          schema:
            type: string
            const: no-store
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: The one error shape at the gateway boundary (spec 0001).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              examples: ["invalid_input"]
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
...
    InvoiceProfile:
      type: object
      additionalProperties: false
      required:
        [legal_name, contact_line, bank_code, bank_name, bank_account_number, bank_account_holder, revision, is_complete, missing_fields, bank_status]
      properties:
        legal_name:
          type: [string, "null"]
        contact_line:
          type: [string, "null"]
        bank_code:
          type: [string, "null"]
        bank_name:
          type: [string, "null"]
        bank_account_number:
          type: [string, "null"]
        bank_account_holder:
          type: [string, "null"]
        revision:
          type: integer
          format: int64
          minimum: 0
        is_complete:
          type: boolean
        missing_fields:
          type: array
          items:
            type: string
            enum: [legal_name, contact_line, bank_code, bank_account_number, bank_account_holder]
        bank_status:
          type: string
          enum: [missing, active, inactive]
    PutInvoiceProfileRequest:
      type: object
      additionalProperties: false
      required:
        [expected_revision, legal_name, contact_line, bank_code, bank_account_number, bank_account_holder]
      properties:
        expected_revision:
          type: integer
          format: int64
          minimum: 0
        legal_name:
          type: [string, "null"]
          maxLength: 120
        contact_line:
          type: [string, "null"]
          maxLength: 200
        bank_code:
          type: [string, "null"]
        bank_account_number:
          type: [string, "null"]
        bank_account_holder:
          type: [string, "null"]
          maxLength: 120
    Bank:
      type: object
      additionalProperties: false
      required: [code, short_name, official_name]
      properties:
        code:
          type: string
          pattern: "^[0-9]{6}$"
        short_name:
          type: string
        official_name:
          type: string
    BankCatalog:
      type: object
      additionalProperties: false
      required: [banks]
      properties:
        banks:
          type: array
          items:
            $ref: "#/components/schemas/Bank"
    RateCurrent:
      type: object
      additionalProperties: false
      required: [rate_amount, currency, effective_from, rate_revision]
      properties:
        rate_amount:
          type: integer
          format: int64
          minimum: 0
          maximum: 1000000000
        currency:
          type: string
          const: VND
        effective_from:
          type: string
          format: date
        rate_revision:
          type: integer
          format: int64
          minimum: 0
    RateAllowedRange:
      type: object
      additionalProperties: false
      required: [from, through]
      properties:
        from:
          type: string
          format: date
        through:
          type: string
          format: date
    ProjectedClassRate:
      type: object
      additionalProperties: false
      required: [effective_from, rate_amount, currency, rate_revision]
      properties:
        effective_from:
          type: string
          format: date
        rate_amount:
          type: integer
          format: int64
          minimum: 0
          maximum: 1000000000
        currency:
          type: string
          const: VND
        rate_revision:
          type: integer
          format: int64
          minimum: 0
    ClassRates:
      type: object
      additionalProperties: false
      required: [class_id, current, allowed_range, archived, rates, projected_revision, history_state]
      properties:
        class_id:
          type: string
          format: uuid
        current:
          $ref: "#/components/schemas/RateCurrent"
        allowed_range:
          $ref: "#/components/schemas/RateAllowedRange"
        archived:
          type: boolean
        rates:
          type: array
          items:
            $ref: "#/components/schemas/ProjectedClassRate"
        projected_revision:
          type: integer
          format: int64
          minimum: 0
        history_state:
          type: string
          enum: [synced, syncing, unavailable]
...
    BillingPeriod:
      type: object
      additionalProperties: false
      required: [year, month]
      properties:
        year:
          type: integer
          minimum: 2000
        month:
          type: integer
          minimum: 1
          maximum: 12
    BillingPeriodDefault:
      type: object
      additionalProperties: false
      required: [server_date, timezone, year, month, minimum_year]
      properties:
        server_date:
          type: string
          format: date
        timezone:
          type: string
        year:
          type: integer
          minimum: 2000
        month:
          type: integer
          minimum: 1
          maximum: 12
        minimum_year:
          type: integer
          const: 2000
    BillingRecoveryDestination:
      type: object
      additionalProperties: false
      required: [route, date, session_id, class_id, rate_date]
      properties:
        route:
          type: string
        date:
          type: [string, "null"]
          format: date
        session_id:
          type: [string, "null"]
          format: uuid
        class_id:
          type: [string, "null"]
          format: uuid
        rate_date:
          type: [string, "null"]
          format: date
    BillingBlocker:
      type: object
      additionalProperties: false
      required: [code, student_id, session_id, class_id, local_date, field, destination]
      properties:
        code:
          type: string
          enum: [profile_incomplete, attendance_incomplete, rate_missing]
        student_id:
          type: [string, "null"]
          format: uuid
        session_id:
          type: [string, "null"]
          format: uuid
        class_id:
          type: [string, "null"]
          format: uuid
        local_date:
          type: [string, "null"]
          format: date
        field:
          type: [string, "null"]
        destination:
          $ref: "#/components/schemas/BillingRecoveryDestination"
    BillingLine:
      type: object
      additionalProperties: false
      required: [session_id, class_id, class_name, local_date, rate_amount, amount, currency]
      properties:
        session_id:
          type: string
          format: uuid
        class_id:
          type: [string, "null"]
          format: uuid
        class_name:
          type: string
        local_date:
          type: string
          format: date
        rate_amount:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
    BillingStudentTotal:
      type: object
      additionalProperties: false
      required: [student_id, student_name, lines, total_amount, currency]
      properties:
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BillingLine"
        total_amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
    IssuedInvoice:
      type: object
      additionalProperties: false
      required:
        [invoice_id, student_id, student_name, invoice_number, total_amount, currency, issued_at, lines]
      properties:
        invoice_id:
          type: string
          format: uuid
        student_id:
          type: string
          format: uuid
        student_name:
          type: string
        invoice_number:
          type: string
        total_amount:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
        issued_at:
          type: string
          format: date-time
        lines:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BillingLine"
    BillingRun:
      type: object
      additionalProperties: false
      required: [billing_run_id, period, generation, created_at, invoices, grand_total, currency]
      properties:
        billing_run_id:
          type: string
          format: uuid
        period:
          $ref: "#/components/schemas/BillingPeriod"
        generation:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/IssuedInvoice"
        grand_total:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
    BillingPeriodState:
      type: object
      additionalProperties: false
      required: [status, period, run]
      properties:
        status:
          type: string
          enum: [unissued, already_issued]
        period:
          $ref: "#/components/schemas/BillingPeriod"
        run:
          oneOf:
            - $ref: "#/components/schemas/BillingRun"
            - type: "null"
    BillingPreview:
      type: object
      additionalProperties: false
      required: [status, period, students, blockers, grand_total, currency, preview_fingerprint, run]
      properties:
        status:
          type: string
          enum: [ready, blocked, empty, already_issued]
        period:
          $ref: "#/components/schemas/BillingPeriod"
        students:
          type: array
          items:
            $ref: "#/components/schemas/BillingStudentTotal"
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/BillingBlocker"
        grand_total:
          type: integer
          format: int64
        currency:
          type: string
          const: VND
        preview_fingerprint:
          type: [string, "null"]
        run:
          oneOf:
            - $ref: "#/components/schemas/BillingRun"
            - type: "null"
    IssueBillingRequest:
      type: object
      additionalProperties: false
      required: [preview_fingerprint]
      properties:
        preview_fingerprint:
          type: string
          minLength: 1
...
security:
  - bearerAuth: []