# services/billing/internal/store/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Attendance struct {
	SessionID  uuid.UUID
	StudentID  uuid.UUID
	TutorID    uuid.UUID
	State      string
	MarkedAt   time.Time
	RecordedAt time.Time
	UpdatedAt  time.Time
}

type BillingRun struct {
	BillingRunID uuid.UUID
	TutorID      uuid.UUID
	PeriodYear   int32
	PeriodMonth  int32
	Generation   int32
	CreatedAt    time.Time
	SupersededAt pgtype.Timestamptz
}

type Class struct {
	ClassID    uuid.UUID
	TutorID    uuid.UUID
	Name       string
	RecordedAt time.Time
	UpdatedAt  time.Time
}

type ClassRate struct {
	ClassID         uuid.UUID
	EffectiveFrom   pgtype.Date
	TutorID         uuid.UUID
	RateAmount      int64
	Currency        string
	RecordedAt      time.Time
	UpdatedAt       time.Time
	RateRevision    int64
	SourcePartition pgtype.Int4
	SourceOffset    pgtype.Int8
}

type ConsumerFailure struct {
	ConsumerName    string
	SourceTopic     string
	SourcePartition int32
	SourceOffset    int64
	EventID         pgtype.UUID
	TutorID         pgtype.UUID
	FailureCategory string
	FailedAt        time.Time
	ResolvedAt      pgtype.Timestamptz
	Resolution      pgtype.Text
	ResolutionCode  pgtype.Text
	ResolvedBy      pgtype.Text
	RepairReference pgtype.Text
}

type ConsumerReadiness struct {
	ConsumerName         string
	ProjectionGeneration uuid.UUID
	State                string
	UpdatedAt            time.Time
}

type ConsumerReplayManifest struct {
	ConsumerName         string
	ProjectionGeneration uuid.UUID
	SourceTopic          string
	TopicIdentity        string
	PartitionSet         []byte
	EarliestOffsets      []byte
	CapturedEndOffsets   []byte
	CompletedOffsets     []byte
	StartedAt            time.Time
	CompletedAt          pgtype.Timestamptz
	OperatorIdentity     string
}

type HandledEvent struct {
	ConsumerName string
	EventID      uuid.UUID
	HandledAt    time.Time
}

type Invoice struct {
	InvoiceID              uuid.UUID
	TutorID                uuid.UUID
	BillingRunID           uuid.UUID
	StudentID              uuid.UUID
	InvoiceNumber          string
	PeriodYear             int32
	PeriodMonth            int32
	TotalAmount            int64
	Currency               string
	IssuedAt               time.Time
	PdfLocation            pgtype.Text
	PaidAt                 pgtype.Timestamptz
	VoidedAt               pgtype.Timestamptz
	VoidReason             pgtype.Text
	ReplacesInvoiceID      pgtype.UUID
	StudentName            string
	PayeeLegalName         string
	PayeeContactLine       string
	PayeeBankName          string
	PayeeBankAccountNumber string
	PayeeBankAccountHolder string
	UpdatedAt              time.Time
}

type InvoiceLine struct {
	InvoiceLineID uuid.UUID
	InvoiceID     uuid.UUID
	TutorID       uuid.UUID
	SessionID     uuid.UUID
	SessionDate   pgtype.Date
	ClassName     string
	RateAmount    int64
	Amount        int64
}

type InvoiceNumberCounter struct {
	TutorID      uuid.UUID
	PeriodYear   int32
	LastSequence int32
	UpdatedAt    time.Time
}

type InvoiceProfile struct {
	TutorID           uuid.UUID
	LegalName         pgtype.Text
	ContactLine       pgtype.Text
	BankName          pgtype.Text
	BankAccountNumber pgtype.Text
	BankAccountHolder pgtype.Text
	CreatedAt         time.Time
	UpdatedAt         time.Time
	BankCode          pgtype.Text
	Revision          int64
	IsComplete        bool
}

type Outbox struct {
	ID           int64
	EventID      uuid.UUID
	EventName    string
	EventVersion int32
	OccurredAt   time.Time
	TutorID      uuid.UUID
	KeyKind      string
	KeyValue     uuid.UUID
	RequestID    string
	Topic        string
	Envelope     []byte
	CreatedAt    time.Time
	PublishedAt  pgtype.Timestamptz
}

type RosterPeriod struct {
	ClassID       uuid.UUID
	StudentID     uuid.UUID
	EffectiveFrom pgtype.Date
	TutorID       uuid.UUID
	EffectiveTo   pgtype.Date
	RecordedAt    time.Time
	UpdatedAt     time.Time
}

type Session struct {
	SessionID   uuid.UUID
	ClassID     uuid.UUID
	TutorID     uuid.UUID
	StartsAt    time.Time
	EndsAt      time.Time
	LocalDate   pgtype.Date
	CancelledAt pgtype.Timestamptz
	RecordedAt  time.Time
	UpdatedAt   time.Time
}

type Student struct {
	StudentID  uuid.UUID
	TutorID    uuid.UUID
	Name       string
	RemovedAt  pgtype.Timestamptz
	RecordedAt time.Time
	UpdatedAt  time.Time
}